@@ -1,6 +1,5 @@
 {
   "name": "react-burger",
-  "homepage": "https://galdenysh.github.io/react-burger/",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -14,12 +13,13 @@
     "@types/react-dom": "^17.0.14",
     "@ya.praktikum/react-developer-burger-ui-components": "^1.13.1",
     "node-sass": "^6.0.0",
-    "react": "^18.0.0",
+    "react": "^18.1.0",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^18.0.0",
-    "react-redux": "^8.0.0",
-    "react-scripts": "5.0.0",
+    "react-dom": "^18.1.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
     "redux": "^4.2.0",
     "redux-thunk": "^2.4.1",
     "sass": "^1.49.11",
