@@ -1,12 +1,12 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { legacy_createStore as createStore, compose, applyMiddleware} from "redux";
-import thunk from 'redux-thunk';
+import { createRoot } from "react-dom/client";
+import { legacy_createStore as createStore, compose, applyMiddleware } from "redux";
+import thunk from "redux-thunk";
 import { Provider } from "react-redux";
-import "./index.css";
+import { BrowserRouter } from "react-router-dom";
 import App from "./components/app/app.jsx";
 import reportWebVitals from "./reportWebVitals";
 import { rootReducer } from "./services/reducers";
+import "./index.css";
 
 declare global {
   interface Window {
@@ -14,22 +14,24 @@ declare global {
   }
 }
 
-const composeEnhancers =
-  typeof window === 'object' && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
-    ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
-    : compose;
+const composeEnhancers = typeof window === "object" && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ : compose;
 
 const enhancer = composeEnhancers(applyMiddleware(thunk));
 
-const store = createStore(rootReducer, enhancer); 
+const store = createStore(rootReducer, enhancer);
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
+const container = document.getElementById("root");
+// @ts-ignore
+const root = createRoot(container!);
+
+root.render(
+  // <React.StrictMode>
+  <Provider store={store}>
+    <BrowserRouter>
       <App />
-    </Provider>  
-  </React.StrictMode>,
-  document.getElementById('root')
+    </BrowserRouter>
+  </Provider>
+  // </React.StrictMode>
 );
 
 reportWebVitals();
