@@ -1,66 +1,107 @@
-import { useEffect, useState, useRef } from "react";
+import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { DndProvider } from "react-dnd";
-import { HTML5Backend } from "react-dnd-html5-backend";
+import { Routes, Route, Navigate, useLocation, useNavigate } from "react-router-dom";
 import AppHeader from "../app-header/app-header.jsx";
-import BurgerIngredients from "../burger-ingredients/burger-ingredients.jsx";
-import BurgerConstructor from "../burger-constructor/burger-constructor.jsx";
+import Login from "../../pages/login/login.jsx";
+import Register from "../../pages/register/register.jsx";
+import ForgotPassword from "../../pages/forgot-password/forgot-password.jsx";
+import ResetPassword from "../../pages/reset-password/reset-password.jsx";
+import Profile from "../../pages/profile/profile.jsx";
+import Ingredients from "../../pages/ingredients/ingredients.jsx";
+import NotFoundPage from "../../pages/not-found-page/not-found-page.jsx";
+import Main from "../../pages/main/main.jsx";
+import ProtectedRoute from "../protected-route/protected-route.jsx";
 import Modal from "../modal/modal.jsx";
-import OrderDetails from "../order-details/order-details.jsx";
 import IngredientDetails from "../ingredient-details/ingredient-details.jsx";
-import styles from "./app.module.scss";
+import Preloader from "../preloader/preloader.jsx";
+import { getCookie } from "../../utils/cookie.js";
+import { getUserData, setRefreshToken } from "../../services/actions/auth.js";
 import { getIngredients } from "../../services/actions/burger.js";
-import { getOrder } from "../../services/actions/order.js";
 
 const App = () => {
-  const [newOrder, setNewOrder] = useState(false);
-  const [visibleOrder, setVisibleOrder] = useState(false);
-  const [visibleIngredient, setVisibleIngredient] = useState(false);
-  const isInitialMount = useRef(true);
-
-  const dispatch = useDispatch();
+  const userData = useSelector((store) => store.authReducer);
   const burderData = useSelector((store) => store.burgerReducer);
-  const orderData = useSelector((store) => store.orderReducer);
+  const location = useLocation();
+  const background = location.state?.background;
+  const navigate = useNavigate();
+  const dispatch = useDispatch();
+
+  const closePopup = () => {
+    navigate(-1);
+  };
 
   useEffect(() => {
+    dispatch(setRefreshToken());
     dispatch(getIngredients());
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
 
-  useEffect(() => {
-    if (isInitialMount.current) {
-      isInitialMount.current = false;
-    } else {
-      dispatch(getOrder(burderData));
-      setVisibleOrder(true);
+    if (getCookie("accessToken")) {
+      dispatch(getUserData());
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [newOrder]);
+  }, []);
 
   return (
     <>
       <AppHeader />
-      <main className={styles.content}>
-        {burderData.isLoading && <p className={`${styles.download} text text_type_main-large`}>Загрузка...</p>}
-        {burderData.hasError && <p className={`${styles.download} text text_type_main-large`}>Произошла ошибка...</p>}
-        {!burderData.isLoading && !burderData.hasError && burderData.ingredientsData.length && (
-          <DndProvider backend={HTML5Backend}>
-            <BurgerIngredients setVisible={setVisibleIngredient} />
-            <BurgerConstructor newOrder={newOrder} setNewOrder={setNewOrder} />
-          </DndProvider>
-        )}
-      </main>
+      <Routes location={background || location}>
+        <Route path="/" element={<Main />} />
+        <Route
+          path="/login"
+          element={
+            <ProtectedRoute anonymous={true}>
+              <Login />
+            </ProtectedRoute>
+          }
+        />
+        <Route
+          path="/register"
+          element={
+            <ProtectedRoute anonymous={true}>
+              <Register />
+            </ProtectedRoute>
+          }
+        />
+        <Route
+          path="/forgot-password"
+          element={
+            <ProtectedRoute anonymous={true}>
+              <ForgotPassword />
+            </ProtectedRoute>
+          }
+        />
+        <Route
+          path="/reset-password"
+          element={
+            <ProtectedRoute anonymous={true}>
+              {userData.resetPasswordAccess ? <ResetPassword /> : <Navigate to="/forgot-password" replace={true} />}
+            </ProtectedRoute>
+          }
+        />
+        <Route
+          path="/profile"
+          element={
+            <ProtectedRoute anonymous={false}>
+              <Profile />
+            </ProtectedRoute>
+          }
+        />
+        <Route path="/ingredients/:id" element={<Ingredients />} />
+        <Route path="*" element={<NotFoundPage />} />
+      </Routes>
 
-      {visibleOrder && !orderData.isLoading && (
-        <Modal setVisible={setVisibleOrder}>
-          {orderData.hasError && <p className={`${styles.download} text text_type_main-large`}>Произошла ошибка...</p>}
-          {!orderData.hasError && <OrderDetails orderNumber={orderData.orderData} />}
-        </Modal>
-      )}
-      {visibleIngredient && (
-        <Modal setVisible={setVisibleIngredient}>
-          <IngredientDetails ingredient={burderData.ingredientSelect} />
-        </Modal>
+      {background && (
+        <Routes>
+          <Route
+            path="/ingredients/:id"
+            element={
+              <Modal closePopup={closePopup}>
+                {burderData.isLoading && <Preloader type={"preloader"} style={{ minHeight: "506px" }} />}
+                {burderData.hasError && <Preloader type={"error"} style={{ minHeight: "506px" }} />}
+                {!burderData.isLoading && !burderData.hasError && burderData.ingredientsData.length && <IngredientDetails />}
+              </Modal>
+            }
+          />
+        </Routes>
       )}
     </>
   );
