@@ -2,10 +2,12 @@ import { useRef } from "react";
 import { useDispatch } from "react-redux";
 import { useDrop, useDrag } from "react-dnd";
 import { ConstructorElement, DragIcon } from "@ya.praktikum/react-developer-burger-ui-components";
+import PropTypes from "prop-types";
 import styles from "./burger-constructor.module.scss";
-import { DECREASE_FILLING_INGREDIENT, REMOVE_FILLING_INGREDIENT } from "../../services/actions/burger.js";
+import { decreaseFillingIngredient, removeFillingIngredient } from "../../services/actions/burger.js";
+import ingredientsPropTypes from "../../utils/types";
 
-const DraggabelConstructorIngredient = (props) => {
+const DraggableConstructorIngredient = (props) => {
   const { ingredient, index, moveIngredient } = props;
   const ref = useRef(null);
   const dispatch = useDispatch();
@@ -51,9 +53,9 @@ const DraggabelConstructorIngredient = (props) => {
     },
   });
 
-  const removeFillingIngredient = (ingredient) => {
-    dispatch({ type: REMOVE_FILLING_INGREDIENT, id: ingredient.constructorId });
-    dispatch({ type: DECREASE_FILLING_INGREDIENT, id: ingredient._id });
+  const remove = (ingredient) => {
+    dispatch(removeFillingIngredient(ingredient.constructorId));
+    dispatch(decreaseFillingIngredient(ingredient._id));
   };
 
   dragRef(dropRef(ref));
@@ -64,7 +66,7 @@ const DraggabelConstructorIngredient = (props) => {
         <DragIcon type="primary" />
       </div>
       <ConstructorElement
-        handleClose={() => removeFillingIngredient(ingredient)}
+        handleClose={() => remove(ingredient)}
         isLocked={false}
         text={ingredient.name}
         price={ingredient.price}
@@ -74,4 +76,10 @@ const DraggabelConstructorIngredient = (props) => {
   );
 };
 
-export default DraggabelConstructorIngredient;
+DraggableConstructorIngredient.propTypes = {
+  ingredient: ingredientsPropTypes.isRequired,
+  index: PropTypes.number.isRequired,
+  moveIngredient: PropTypes.func.isRequired,
+};
+
+export default DraggableConstructorIngredient;
