@@ -1,11 +1,13 @@
-import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useDrop } from "react-dnd";
 import { ConstructorElement } from "@ya.praktikum/react-developer-burger-ui-components";
 import { v4 as uuidv4 } from "uuid";
-import DraggabelConstructorIngredient from "./darggable-constructor-ingredient.jsx";
+import DraggableConstructorIngredient from "./darggable-constructor-ingredient.jsx";
+import PropTypes from "prop-types";
 import styles from "./burger-constructor.module.scss";
-import { ADD_BUN_INGREDIENT, ADD_FILLING_INGREDIENT, INCREASE_FILLING_INGREDIENT, SET_FILLING_INGREDIENT } from "../../services/actions/burger.js";
+import { addBunIngredient, addFillingIngredient, increaseFillingIngredient, setFillingIngredient } from "../../services/actions/burger.js";
+import ingredientsPropTypes from "../../utils/types.js";
+import bunImage from "../../images/bun.png";
 
 const DropTargetIngredients = (props) => {
   const { bunSelect, fillingSelect } = props;
@@ -17,19 +19,14 @@ const DropTargetIngredients = (props) => {
   const dispatch = useDispatch();
   const ingredientsData = useSelector((store) => store.burgerReducer.ingredientsData);
 
-  useEffect(() => {
-    dispatch({ type: ADD_BUN_INGREDIENT, payload: bunSelect });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
   const onDropHandler = (itemId) => {
     const ingredientTarget = ingredientsData.filter((ingredient) => itemId === ingredient._id);
     const ingredientTargetWithId = { ...ingredientTarget[0], constructorId: uuidv4() };
 
     // eslint-disable-next-line no-unused-expressions
     ingredientTargetWithId.type === "bun"
-      ? dispatch({ type: ADD_BUN_INGREDIENT, payload: ingredientTargetWithId })
-      : (dispatch({ type: ADD_FILLING_INGREDIENT, payload: ingredientTargetWithId }), dispatch({ type: INCREASE_FILLING_INGREDIENT, id: itemId }));
+      ? dispatch(addBunIngredient(ingredientTargetWithId))
+      : (dispatch(addFillingIngredient(ingredientTargetWithId)), dispatch(increaseFillingIngredient(itemId)));
   };
 
   const moveIngredient = (dragIndex, hoverIndex) => {
@@ -37,32 +34,49 @@ const DropTargetIngredients = (props) => {
     const dragItem = newFillingSelect.splice(dragIndex, 1);
     newFillingSelect.splice(hoverIndex, 0, dragItem[0]);
 
-    dispatch({ type: SET_FILLING_INGREDIENT, payload: newFillingSelect });
+    dispatch(setFillingIngredient(newFillingSelect));
   };
 
   return (
     <div className={styles.elements} ref={dropTarget}>
       <div className={`${styles.ingredientElement} ml-2`}>
-        <ConstructorElement type="top" isLocked={true} text={`${bunSelect.name} (верх)`} price={bunSelect.price} thumbnail={bunSelect.image} />
+        {Object.keys(bunSelect).length !== 0 && (
+          <ConstructorElement type="top" isLocked={true} text={`${bunSelect.name} (верх)`} price={bunSelect.price} thumbnail={bunSelect.image} />
+        )}
+        {Object.keys(bunSelect).length === 0 && (
+          <ConstructorElement type="top" isLocked={true} text="Пожалуйста, перенесите сюда булку для создания заказа" price={0} thumbnail={bunImage} />
+        )}
       </div>
       <ul className={`${styles.ingredientsList} pr-2`}>
         {!fillingSelect.length && (
           <>
-            <p className={`${styles.emptyList} text text_type_main-default text_color_inactive`}>Здесь пусто.</p>
-            <p className={`${styles.emptyList} text text_type_main-default text_color_inactive`}>
+            <p className={`${styles.emptyList} text text_type_main-default text_color_inactive`} style={{ opacity: "0.4" }}>
+              Здесь пусто.
+            </p>
+            <p className={`${styles.emptyList} text text_type_main-default text_color_inactive`} style={{ opacity: "0.4" }}>
               Вы можете добавить ингредиенты в список, перетащив их карточку из корзины сюда.
             </p>
           </>
         )}
         {fillingSelect.map((ingredient, index) => (
-          <DraggabelConstructorIngredient ingredient={ingredient} index={index} moveIngredient={moveIngredient} key={ingredient.constructorId} />
+          <DraggableConstructorIngredient ingredient={ingredient} index={index} moveIngredient={moveIngredient} key={ingredient.constructorId} />
         ))}
       </ul>
       <div className={`${styles.ingredientElement} ml-2`}>
-        <ConstructorElement type="bottom" isLocked={true} text={`${bunSelect.name} (низ)`} price={bunSelect.price} thumbnail={bunSelect.image} />
+        {Object.keys(bunSelect).length !== 0 && (
+          <ConstructorElement type="bottom" isLocked={true} text={`${bunSelect.name} (низ)`} price={bunSelect.price} thumbnail={bunSelect.image} />
+        )}
+        {Object.keys(bunSelect).length === 0 && (
+          <ConstructorElement type="bottom" isLocked={true} text="Пожалуйста, перенесите сюда булку для создания заказа" price={0} thumbnail={bunImage} />
+        )}
       </div>
     </div>
   );
+};
+
+DropTargetIngredients.propTypes = {
+  bunSelect: PropTypes.object.isRequired,
+  fillingSelect: PropTypes.arrayOf(ingredientsPropTypes),
 };
 
 export default DropTargetIngredients;
