@@ -1,24 +1,27 @@
-import { useDispatch } from "react-redux";
 import { useDrag } from "react-dnd";
 import { Counter, CurrencyIcon } from "@ya.praktikum/react-developer-burger-ui-components";
-import { ADD_INFO_INGREDIENT } from "../../services/actions/burger.js";
 import styles from "./burger-ingredients.module.scss";
+import { useLocation, useNavigate } from "react-router-dom";
+import ingredientsPropTypes from "../../utils/types";
 
 const DraggableIngredient = (props) => {
-  const { ingredient, setVisible } = props;
-  const dispatch = useDispatch();
-  const [, dragRef] = useDrag({
+  const { ingredient } = props;
+  const navigate = useNavigate();
+  const location = useLocation();
+  const [{ isDragging }, dragRef] = useDrag({
     type: "ingredient",
     item: { id: ingredient._id },
+    collect: (monitor) => ({
+      isDragging: monitor.isDragging(),
+    }),
   });
 
-  const openPopup = (ingredient) => {
-    setVisible(true);
-    dispatch({ type: ADD_INFO_INGREDIENT, payload: ingredient });
+  const openPopup = () => {
+    navigate(`/ingredients/${ingredient._id}`, { state: { background: location } });
   };
 
   return (
-    <li className={`${styles.ingredientsItem}`} onClick={() => openPopup(ingredient)} ref={dragRef}>
+    <li className={`${styles.ingredientsItem}`} style={{ cursor: isDragging ? "grabbing" : "grab" }} onClick={openPopup} ref={dragRef}>
       {ingredient.qty > 0 && <Counter count={ingredient.qty} size="default" />}
       <img src={ingredient.image} alt={ingredient.name} />
       <div className={styles.price}>
@@ -28,6 +31,10 @@ const DraggableIngredient = (props) => {
       <p className="text text_type_main-default mt-2">{ingredient.name}</p>
     </li>
   );
+};
+
+DraggableIngredient.propTypes = {
+  ingredient: ingredientsPropTypes.isRequired,
 };
 
 export default DraggableIngredient;
