@@ -1,21 +1,14 @@
 import { useEffect } from "react";
-import { useDispatch } from "react-redux";
 import ReactDOM from "react-dom";
 import PropTypes from "prop-types";
 import { CloseIcon } from "@ya.praktikum/react-developer-burger-ui-components";
 import ModalOverlay from "../modal-overlay/modal-overlay";
-import { REMOVE_INFO_INGREDIENT } from "../../services/actions/burger";
 import styles from "./modal.module.scss";
 
 const modalRoot = document.getElementById("react-modals");
 
 const Modal = (props) => {
-  const dispatch = useDispatch();
-
-  const closePopup = () => {
-    props.setVisible(false);
-    dispatch({ type: REMOVE_INFO_INGREDIENT });
-  };
+  const { closePopup, children } = props;
 
   const closePopupEsc = (evt) => {
     if (evt.key === "Escape") {
@@ -35,7 +28,7 @@ const Modal = (props) => {
   return ReactDOM.createPortal(
     <section className={styles.popup}>
       <div className={styles.container}>
-        {props.children}
+        {children}
         <div className={`${styles.closeBtn} mt-15 mr-10`} onClick={closePopup}>
           <CloseIcon type="primary" />
         </div>
@@ -47,8 +40,8 @@ const Modal = (props) => {
 };
 
 Modal.propTypes = {
-  children: PropTypes.object.isRequired,
-  setVisible: PropTypes.func.isRequired,
+  children: PropTypes.oneOfType([PropTypes.element.isRequired, PropTypes.array.isRequired]),
+  closePopup: PropTypes.func.isRequired,
 };
 
 export default Modal;
